requirements.txt
anyio==3.7.0
asyncpg==0.27.0
certifi==2023.5.7
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.97.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
orjson==3.9.1
packaging==23.1
pip==21.3.1
pluggy==1.0.0
pydantic==1.10.9
pytest==7.3.2
pytest-asyncio==0.21.0
python-dotenv==1.0.0
python-multipart==0.0.6
setuptools==60.2.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing-extensions==4.6.3
uvicorn==0.22.0
wheel==0.37.1
SQLAlchem==2.0.16